:truck: Moves pieces to own files
