+++ packages/reactivity/src/index.ts
@@ -1,221 +1,26 @@
+import { Effect, effect, release, stop } from './Effect';
+import type { ReactiveEffect } from './Effect';
+import { Signal, untrack } from './Signal';
 import { nextTick } from './nextTick';
+import { $PROXY, $RAW, reactive, toRaw } from './reactive';
+
+export {
+  Signal,
+  untrack,
+  Effect,
+  ReactiveEffect,
+  effect,
+  release,
+  stop,
+  reactive,
+  toRaw,
+};
 
 declare global {
   interface ImportMeta {
     DEBUG: boolean;
   }
 }
-
-let nextId = 1;
-export class Signal<T = unknown> {
-  id: number;
-  value: T;
-  dependents: Set<Effect> = new Set();
-  constructor(value: T) {
-    this.id = nextId++;
-    import.meta.DEBUG &&
-      console.log(
-        'creating Signal:',
-        typeof value === 'object' ? 'object' : value,
-        this.id,
-      );
-    this.value = value;
-  }
-  get() {
-    import.meta.DEBUG &&
-      console.log(
-        'getting Signal:',
-        typeof this.value === 'object' ? 'object' : this.value,
-        this.id,
-      );
-    if (dontTrack) return this.value;
-    const activeEffect = effectStack.at(-1);
-    if (activeEffect) {
-      import.meta.DEBUG &&
-        console.log(
-          'registering signal:',
-          this.id,
-          ', to effect:',
-          activeEffect.id,
-        );
-      activeEffect.register(this);
-      this.dependents.add(activeEffect);
-    }
-    return this.value;
-  }
-  set(value: T) {
-    import.meta.DEBUG && console.log('setting signal:', value, this.id);
-    this.value = value;
-    this.dependents.forEach((effect) => effect.rerun());
-  }
-  release(effect: Effect) {
-    import.meta.DEBUG &&
-      console.log('releasing effect:', effect.id, ', from signal:', this.id);
-    this.dependents.delete(effect);
-  }
-}
-
-let dontTrack = false;
-export const untrack = <T>(cb: () => T): T => {
-  dontTrack = true;
-  const output = cb();
-  dontTrack = false;
-  return output;
-};
-const effectStack: Effect[] = [];
-const effectQueue: Effect[] = [];
-let waitingForEffects = false;
-const addToEffectQueue = (effect: Effect) => {
-  const position = effectQueue.findIndex((e) => e.id > effect.id);
-  if (position > -1) effectQueue.splice(position, 0, effect);
-  else effectQueue.push(effect);
-  if (waitingForEffects) return;
-  waitingForEffects = true;
-  queueMicrotask(processEffects);
-};
-
-const processEffects = () => {
-  while (effectQueue.length) effectQueue.shift()?.run();
-  waitingForEffects = false;
-};
-
-export class Effect {
-  id: number;
-  operation: () => void;
-  dependencies: Set<Signal> = new Set();
-  constructor(cb: () => void) {
-    this.id = nextId++;
-    import.meta.DEBUG && console.log('creating effect:', this.id);
-    this.operation = cb;
-    this.run();
-  }
-  run() {
-    effectStack.push(this);
-    import.meta.DEBUG && console.log('running effect:', this.id);
-    this.operation();
-    effectStack.pop();
-  }
-  register(signal: Signal) {
-    import.meta.DEBUG &&
-      console.log(
-        'registering signal:',
-        signal.id,
-        ', to this effect:',
-        this.id,
-      );
-    this.dependencies.add(signal);
-  }
-  rerun() {
-    import.meta.DEBUG && console.log('triggering rerun on effect:', this.id);
-    if (this === effectStack.at(-1)) return;
-    this.release();
-    addToEffectQueue(this);
-  }
-  release(cleanUp?: boolean) {
-    import.meta.DEBUG && console.log('releasing effect', this.id);
-    this.dependencies.forEach((signal) => {
-      signal.release(this);
-      this.dependencies.delete(signal);
-    });
-    if (cleanUp) this.operation = () => {};
-  }
-  dequeue() {
-    const idx = effectQueue.indexOf(this);
-    if (idx > -1) effectQueue.splice(idx, 1);
-  }
-}
-
-const $PROXY = Symbol('$PROXY');
-const $RAW = Symbol('$RAW');
-const proxyMap = new WeakMap<object, object>();
-const reactiveNodes = new WeakMap<object, { [key: string | symbol]: Signal }>();
-// eslint-disable-next-line @typescript-eslint/ban-types
-const wrappableObjects: (Function | undefined)[] = [Array, Object, undefined];
-const isWrappable = (obj: object) => wrappableObjects.includes(obj.constructor);
-
-export const reactive = <T extends object>(obj: T): T => {
-  if (typeof obj !== 'object' || obj === null || !isWrappable(obj)) return obj;
-  if (proxyMap.has(obj)) return proxyMap.get(obj) as T;
-  if ($RAW in obj) return obj;
-  reactiveNodes.set(obj, Object.create(null));
-  const wrapped = new Proxy(obj, {
-    has(target, p) {
-      if (p === $RAW) return true;
-      if (p === $PROXY) return true;
-      return Reflect.has(target, p);
-    },
-    get(target, p: string | symbol, reciever) {
-      if (p === $RAW) return target;
-      if (p === $PROXY) return proxyMap.get(target);
-      if (!(p in target)) return undefined;
-      if (!Object.hasOwnProperty.call(target, p)) return Reflect.get(target, p);
-      const descriptor = Object.getOwnPropertyDescriptor(target, p);
-      if (descriptor?.get) return descriptor.get.call(reciever);
-      const nodes = reactiveNodes.get(target);
-      if (!nodes) return undefined;
-      if (p in nodes) return nodes[p].get();
-      const signal = wrap(Reflect.get(target, p));
-      nodes[p] = signal;
-      return signal.get();
-    },
-    set(target, p: string | symbol, newValue, reciever) {
-      if (import.meta.DEBUG)
-        console.log(
-          'setting',
-          p,
-          'to',
-          newValue,
-          Reflect.get(target, p) === newValue ? 'SAME' : 'DIFFERENT',
-        );
-      const nodes = reactiveNodes.get(target);
-      if (!nodes) return false;
-      const descriptor = Object.getOwnPropertyDescriptor(target, p);
-      if (descriptor?.set) {
-        descriptor.set.call(reciever, newValue);
-        return true;
-      }
-      if (p in nodes) {
-        if (untrack(() => nodes[p].get()) === newValue) return true;
-        nodes[p].set(
-          typeof newValue === 'object' ? reactive(newValue) : newValue,
-        );
-      } else {
-        const signal = wrap(Reflect.get(target, p));
-        nodes[p] = signal;
-        signal.set(
-          typeof newValue === 'object' ? reactive(newValue) : newValue,
-        );
-      }
-      Reflect.set(target, p, newValue);
-      return true;
-    },
-  });
-  Object.defineProperty(obj, $PROXY, {
-    enumerable: false,
-    configurable: false,
-    value: wrapped,
-  });
-  proxyMap.set(obj, wrapped);
-
-  return wrapped as T;
-};
-
-// exports @vue/reactivity API
-export const effect = (cb: () => void) => new Effect(cb);
-export const release = (effect: Effect) => effect.release(true);
-export const stop = release;
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export const toRaw = <T extends object>(obj: T): T => Reflect.get(obj, $RAW);
-export type ReactiveEffect = Effect;
-
-const wrap = <T>(item: T): Signal<T> => {
-  if (item instanceof Signal) return item;
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  if (typeof item === 'object' && item !== null)
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    return new Signal(reactive(item as any));
-  return new Signal(item);
-};
 
 if (import.meta.vitest) {
   describe('reactivity', () => {
@@ -429,6 +234,13 @@ if (import.meta.vitest) {
       // eslint-disable-next-line @typescript-eslint/ban-ts-comment
       // @ts-ignore
       expect(proxy[$PROXY]).toBe(proxy);
+    });
+    it('allows accessing the raw object from the $RAW property', () => {
+      const obj = { value: 42 };
+      const proxy = reactive(obj);
+      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+      // @ts-ignore
+      expect(proxy[$RAW]).toBe(obj);
     });
     it('returns the same proxy when wrapping the same target', () => {
       const obj = { value: 42 };
