+++ size.json
@@ -1,22 +1,22 @@
 {
   "reactivity": {
     "minified": {
-      "pretty": "2.15 kB",
-      "raw": 2146
+      "pretty": "2.16 kB",
+      "raw": 2161
     },
     "brotli": {
-      "pretty": "901 B",
-      "raw": 901
+      "pretty": "904 B",
+      "raw": 904
     }
   },
   "alpinejs": {
     "minified": {
       "pretty": "33.4 kB",
-      "raw": 33356
+      "raw": 33373
     },
     "brotli": {
       "pretty": "11.3 kB",
-      "raw": 11320
+      "raw": 11275
     }
   }
 }
